requirements/common.txt
asgiref==3.2.10
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
djangorestframework-3.12.1
dj-database-url-0.5.0
django-heroku-0.3.1
psycopg2-2.8.6
whitenoise-5.2.0
gunicorn-20.0.4
